+++ mutraction-dom-docs/src/forEach.tsx
@@ -48,6 +48,13 @@ export function forEach() {
                         <li><code>index</code> is the array index of the item.</li>
                         <li><code>array</code> is the array of items.</li>
                     </ul>
+                    <p>
+                        This function can either return a DOM node or an options object.
+                    </p>
+                    <ul>
+                        <li><code>node</code> is the output node.</li>
+                        <li><code>cleanup</code> is an optional callback to invoke when replacing the previous node.</li>
+                    </ul>
                 </li>
             </ul>