+++ CHANGELOG.md
@@ -11,5 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- ErrorBoundary
-- Swapper
+- `ErrorBoundary`
+- `Swapper`
+
+### Changed
+
+- Added cleanup callback option to `ForEach`